note of the author (David Cantrell) of the fix:
The attached version of lock.t fixes this.  I *think* that it was a bug
in the tests, and it was confused about which was the child and which
the parent process after forking.

Note that forking then confuses Test::More as it sees the wrong number
of tests even though the TAP output is correct.  I've fixed that too by
taking out Test::More and doing all the output with print statements.
This also fixes it on Solaris, and I would expect it to fix the problem
on HPUX and Cygwin too, so you may be able to get rid of the skip
entirely.

But please, someone who understands what IO::All is meant to do, check
that what I've done is correct.  The "fix" is only a guess!