Add a fix for io->file("foobar")->assert.
This is a fix for http://bugs.debian.org/cgi-bin/bugreport.cgi?bug=733680 .

$ ls -l; echo foo | perl -MIO::All -e 'io("-") > io->file("./bar")->assert'; ls -l
total 0
Can't open file './bar' for output:
Is a directory at -e line 1.
total 0
drwxr-xr-x 1 abe  abe      0 Dec 30 23:51 bar/
$

The culprit appears to be that «-d ""» no longer returns true - it may be
a regression in glibc or the Linux kernel. Anyway, this commit
worksaround it, along with a test. I was able to reproduce it on very
old versions of IO::All so I don't know why it surfaced now.